+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `ConfiguredRadium` HOC to use Radium with custom configuration.
+- Fix: Update Radium version and add a custom plugin to fix `prefixer` data.
+- Fix: Fix `Carousel` prerender.
 - Fix: Update styles on `Tour`, `Popover` and `Modal` components.
 - Feature: Add `k-ButtonIcon--cross--tiny` class.